refactor(worktree): extract lock primitive and dedup the create/enter flows
Cleanup pass over the worktree provisioning/enter/lock feature — no
behavior change:

- Move the pure file-backed cross-agent lock (acquire/release, ownership,
  liveness) into src/utils/worktree-lock.ts; the tool keeps only the
  session-aware orchestration. Reusable by a future exit/remove tool.
- Reuse @/utils/fs writeJsonFile/readJsonFile for the lock file (adds
  flush + auto-mkdir) instead of hand-rolled stringify/read.
- Extract resolveWorktreeContext for the cwd/repoRoot/primaryRoot/managedDir
  prelude shared by the create and enter flows.
- Skip the previous-cwd lock-release git spawn when leaving a path outside
  .letta/worktrees/ (the common switch-in-from-main case).
- symlinkDirIntoWorktree returns {linked, note}; linkGitHooks drops its
  duplicate lstat and the .startsWith("symlinked") return-string sniff.
- Drop the orphaned ProjectSettings.worktree field (read directly from disk).
